+++ app/src/main/java/com/example/calculatrice/Multiply.java
@@ -5,7 +5,7 @@ import android.widget.TextView;
 public class Multiply extends AChain {
 
     @Override
-    public void calculate(int num1, int num2, String operation, TextView tw) {
+    public void calculate(float num1, float num2, String operation, TextView tw) {
 
         if(operation.compareTo("*") == 0){
             result = num1 * num2;
